% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LadjSp.R
\name{LadjSp}
\alias{LadjSp}
\title{Ladjsp}
\usage{
LadjSp(ARL0nom, m, n)
}
\arguments{
\item{ARL0nom}{valor da distribuição}

\item{m}{valor das quantidades}

\item{n}{quantidade}
}
\description{
Utilizada para cálculos probaílisticos
}
\examples{
LadjSp(5,17,3)

ARL0XbarSp
}
